Strip Metadata down to the minimum, put more responsibility on Timestamps.
