+++ lib/resque/plugins/meta/metadata.rb
@@ -63,73 +63,10 @@ module Resque
           data[key.to_s] = val
         end
 
-        def start!
-          @started_at = Time.now
-          self['started_at'] = to_time_format_str(@started_at)
-          save
-        end
-
-        def started_at
-          @started_at ||= from_time_format_str('started_at')
-        end
-
-        def finish!
-          data['succeeded'] = true unless data.has_key?('succeeded')
-          @finished_at = Time.now
-          self['finished_at'] = to_time_format_str(@finished_at)
-          save
-        end
-
-        def finished_at
-          @finished_at ||= from_time_format_str('finished_at')
-        end
-
-        def expire_at
-          if finished? && expire_in > 0
-            finished_at.to_i + expire_in
-          else
-            0
-          end
-        end
-
-        def enqueued?
-          !started?
-        end
-
-        def working?
-          started? && !finished?
-        end
-
-        def started?
-          !!started_at
-        end
-
-        def finished?
-          !!finished_at
-        end
-
-        def fail!
-          self['succeeded'] = false
-          finish!
-        end
-
-        def succeeded?
-          finished? ? self['succeeded'] : nil
-        end
-
-        def failed?
-          finished? ? !self['succeeded'] : nil
-        end
-
-        def seconds_enqueued
-          (started_at || Time.now).to_f - enqueued_at.to_f
-        end
-
-        def seconds_processing
-          if started?
-            (finished_at || Time.now).to_f - started_at.to_f
-          else
-            0
+        # methods in modules can be easily overridden or extended later (with more modules)
+        include module InstanceMethods
+          def expire_at
+            Time.now.to_i + expire_in
           end
         end
